requirements.txt
streamlit==1.22.0
pandas==2.0.0
gspread==5.9.0
oauth2client==4.1.3
langchain==0.0.179
openai==0.27.7
tabulate==0.9.0
python-environ==0.4.54
retrying==1.3.4